Reintegrate pnpm engine (#1203)
* This builds on top of #1199. The goal of this PR is to:
- eliminate the contents of the former `pnpm-engine` CNB found in `src/pnpm_engine`
- leverage the functions from `util::npm_registry` for downloading and installing packages from the registry to replace the previous Corepack-based functionality

This simplified process for package manager installation has laid the groundwork for moving away from using the `corepack` tooling directly while still supporting the use of the `packageManager` field in `package.json`. This is especially important as Corepack won't be officially supported/bundled with Node.js in the future.
Because of this fundamental change, the integration test snapshots do show both layer name and buildpack output changes.

* Reintegrate yarn engine (#1204)

* Support installation of Yarn via `yarn.engines` and `packageManager` from `package.json`

This uses the same packument installation process laid out in earlier commits which already handles `npm` and `pnpm` installs. This works without modifications to the `src/yarn` CNB code due to how it checks if Yarn is installed and skips its own internal install process in that case.

With this change, all Corepack-related code can finally be removed.

* Support installation of default Yarn if `yarn.lock` exists

This detects an extra case from the Yarn CNB which will default to using `yarn@1.22.x` if `yarn.lock` exists but no `engines.yarn` or `packageManager` field is found.

With this change, the installation code from the prior Yarn CNB can be removed.

* Detect vendored Yarn installation

The prior Yarn CNB would bootstrap Yarn 1.x to deal with projects using a vendored Yarn distribution. This isn't necessary as the file referenced in the `yarnPath` of `.yarnrc.yml` should be a fully functional `yarn` binary. This PR introduces a new uncached layer for this case that creates a symlink from the `yarnPath` to `yarn` and makes this available at both build and launch.

* Removed "Yarn operating in CLI mode {x}" output

Now that we no longer install a bootstrap version of Yarn to handle vendored Yarn configurations, this extra output line is unnecessary. The version installed will always match the version reported in this removed output.

* Lift package manager installation step up

Now that the code from the CNBs that were responsible for package manager installation has been removed, it's possible to extract the package manager installation process up to a standalone step in the overall buildpack execution process.

* Update CHANGELOG.md

* Unit tests for `Yarnrc` functionality and error snapshots

* Reintegrate npm install (#1205)

* Switch to use `BuildpackConfig` for `npm_install`

The `buildpack_config` module is intended to replace the individual handling of buildplans and project.toml as configuration sources. Eventually this module will replace `utils::buildplan` and `utils::config` but the old `pnpm_install` and `yarn` CNB flows still rely on them.

* Change `determine_package_manager` to return a non-optional value

Since npm is bundled with Node.js, we can default to that when no package manager information is provided.

* Chain runtime installation flow

This mirrors the package manager install flow

* Remove `npm_install` detect

* Lift dependency installation from `npm_install` into top-level

* Lift build script execution from `npm_install` into top-level

* Lift dev dependency pruning from `npm_install` into top-level

* Move npm runtime configuration into an exec.d script

* Lift process registration from `npm_install` into top-level

* Merge npm-related integration tests

* Reintegrate yarn (#1206)

* Remove `yarn` detect

* Switch to use `BuildpackConfig` for `yarn`

* Lift global cache setup from `yarn` into top-level

* Lift dependency installation from `yarn` into top-level

For consistency with the npm cache layer, the Yarn cache layer has been renamed to `yarn_cache`. This can be seen in changes to the integration test snapshots.

* Lift build script execution from `yarn` into top-level

* Lift dev dependency pruning from `yarn` into top-level

* Lift process registration from `yarn` into top-level

This final step allows the `src/yarn` directory to finally be removed.

* Reintegrate pnpm install (#1208)

* Remove `pnpm_install` detect

* Switch to use `BuildpackConfig` for `pnpm_install`

* Lift dependency installation from `pnpm_install` into top-level

* Add pnpm persistent store cleanup

This feature requires the use of the CNB store to keep a counter of how many builds the cache has been around for.

* Lift build script execution from `pnpm_install` into top-level

* Lift dev dependency pruning from `pnpm_install` into top-level

* Lift process registration from `pnpm_install` into top-level

This final step allows the `src/pnpm_install` directory to finally be removed.

* Make `ErrorMessage` the primary Err-type for the buildpack

* Merge pnpm integration tests