Replace vendored yarn link with executable (#1224)
* Replace vendored `yarn` link with executable wrapper

When a vendored Yarn script is configured in the `yarnPath` of `.yarnrc.yml` (e.g.; `yarnPath: .yarn/releases/yarn-4.0.0.cjs`), we have three options:
1. Install a `yarn` binary (defaults to `1.22.x`) to act as a wrapper that, as part of the Yarn CLI behavior, delegates execution to the vendored script.
2. Write a symbolic link from `yarn` to `.yarn/releases/yarn-4.0.0.cjs` that invokes the script directly.
3. Add a custom `yarn` wrapper that delegates execution to the vendored script.

Initially, this CNB used method 1 but performing a full Yarn install just to delegate to the intended vendored Yarn seemed inefficient so, as part of #1204, we switched to method 2 to simplify the process. Turns out this method is less than ideal as:
 - the vendored Yarn script may not have the required executable permissions (#1215)
 - we may not have permissions to force executable permissions (#1221) as the vendored script resides in the `/workspace` directory and buildpacks should treat those files as immutable

This leaves method 3, implemented in this PR, which creates the following `yarn` executable:

```sh
#!/usr/bin/env node
require('/path/to/vendored-yarn.cjs');
```

The executable permissions can be guaranteed for this wrapper binary as it's created in the `yarn_vendored` layer owned by this buildpack. The vendored Yarn script is written as an IIFE so requiring the file is enough to execute it.

* Fix clippy errors

* Update CHANGELOG.md

* Use serde::json to interpolate value into script

Co-authored-by: David Zuelke <dzuelke@salesforce.com>
Signed-off-by: Colin Casey <casey.colin@gmail.com>

---------

Signed-off-by: Colin Casey <casey.colin@gmail.com>
Co-authored-by: David Zuelke <dzuelke@salesforce.com>