Capture user requested value for versions
When this buildpack reads the value entered by the user for a requested version of:
- Node.js
- npm
- pnpm
- Yarn

This value is parsed as a `Requirement(Range)`. Because the current implementation of `Requirement` implements `Display` to print the parsed `Range` representation of the given value, it doesn't match what the user entered (e.g.; `24.x` → `>=24.0.0 <25.0.0`).

This PR changes this behavior to retain the user value and report that in buildpack output and instrumentation instead of the internal `Range` representation. Snapshots changes will highlight the difference more clearly.